requirements.txt
numpy
optuna
pandas
scikit-learn
scipy
statsmodels
tqdm
joblib
torch
biopython
fair-esm